Update mainstream_slug_updater to use a specified user.
Instead of creating a fake user.  This fake user was being created
without a uid etc, and not linked to Signon, which isn't great.  Now
that we've introduced a unique index on User.uid, this has been causing
problems.

This also includes a migration to set a uid and disable the fake user.