Disable a non-deterministically failing assertion
This assertion was failing frequently, but not 100% of the time, in our
CI system, and blocking deploys.  We are pretty confident that this is a
problem with our tests, rather than the actual system.

We've spent considerable time trying to make the test be reliable (see
3a53f28348fe7120c2aaa2d1391ea6aa73c63d66), but have failed.

Removing a test like this makes me very sad, but tests which continually
give false positives aren't useful.